lsp-rust: add semantic token moodifiers
This is more of a proof-of-concept than anything else.

It works for the "mutable" modifier, adding an underline. This matches
what the VsCode client does, but clashes with the lsp-mode default for
read vs write highlighted variables. The rest have a no-operation face
assigned, but need to be there to get the numbers right.

It is based on the approach in lsp-terraform.el.  I tried setting it
in the client registration as per lsp-clojure.el, but the final
mapping of fonts to indexes does not work for some reason,
as the call to lsp--semantic-tokens-initialize-workspace does not get
the correct value from (lsp-semantic-tokens--modifier-faces-for
client), despite them being set elsewhere. I suspect a timing of
update problem.