semantic-tokens: cache modifier lookups
Make a buffer-local cache of the mapping from a modifier bitmap to the
set of faces it resolves to.

The assumption is that for a given language the modifier usage will
tend to cluster, so the lookup table will not be large.

(cherry picked from commit 804e886753f23f32169389bfbc3087293e792c82)